+++ src/main/java/finalmission/controller/dto/RoomCreateResponse.java
@@ -1,4 +1,4 @@
-package finalmission.controller;
+package finalmission.controller.dto;
 
 import finalmission.domain.Room;
 import java.time.LocalDate;
